(SERVER-860) Rename RegisteredLinkedlockingDeque to JRubyPool
